Coded Initial Project
This is kind of a "rest of the owl" commit.  It only took me a couple days/sessions to code this so I put my head down and threw this together.  There are some issues which I will clean up shortly:
Not paying attention to my initial commit, I installed in developer mode and accidentally included a .egg-info folder and a .pyc compiled file.  A little oopsy which I'll just untrack once this is in.