Added in algorithm for overlapping nodes and edges
This was not the best algorithm, I ended up doing a couple iterations of "what looks visually acceptable."  A better algorithm should be come up with next; however, the focus of these commits was to create the diagram which was successful.  My thoughts are along the lines of calculating the longest distance to a node and set this distance as the x coordinate.  This will handle the "flow" of work well.  Next for y feels like an optimization where I should attempt to minimize the absolute sum of vertical deviances from the attaching nodes constraining against overlaps of nodes or edges.  This fundamentally will minimize to 0 if possible.  It could get very complex if I handle the algorithm globally (imagine 30 departments with project plans all meeting together for a combined objective).  I bet I could handle this left to right, locally optimizing for each x coordinate.  I'm not sure if this will work for all cases, but I think it would be a good start which won't be too algorithmically intensive.