Added some utility functions around file picking and conversion
Can now read an Excel file and output a pipe delimited txt file
Can now run the main function which invokes a file picker
Really feeling the need for static scans on the incoming data to make sure it has the correct shape.  It's tough to debug visually, but running some simple code to check for repeated object names would go a long way in clearing this up.